+++ src/Hx.Abp.Attachment.Application.Contracts/Hx/Abp/Attachment/Application/Contracts/HxAbpAttachmentApplicationContractsModule.cs
@@ -1,3 +1,4 @@
+using Hx.Abp.Attachment.Domain.Shared;
 using Volo.Abp.Application;
 using Volo.Abp.Authorization;
 using Volo.Abp.Modularity;
@@ -6,7 +7,8 @@ namespace Hx.Abp.Attachment.Application.Contracts
 {
     [DependsOn(
         typeof(AbpAuthorizationModule),
-        typeof(AbpDddApplicationModule)
+        typeof(AbpDddApplicationModule),
+        typeof(HxAbpAttachmentDomainSharedModule)
         )]
     public class HxAbpAttachmentApplicationContractsModule : AbpModule
     {