+++ src/Hx.Abp.Attachment.Dmain.Shared/Hx/Abp/Attachment/Domain/Shared/AttachmentLocalizationResource.cs
@@ -1,0 +1,13 @@
+using Volo.Abp.Localization;
+
+namespace Hx.Abp.Attachment.Domain.Shared
+{
+    /// <summary>
+    /// 附件模块本地化资源
+    /// </summary>
+    [LocalizationResourceName("Attachment")]
+    public class AttachmentLocalizationResource
+    {
+    }
+}
+